consolidated dist and toolkit
